sdk: move `storage` to `wallet` module
